Channel flood / reordering and dither support when using OpenColorIO (#838)
### Summarize your change.

If the OpenColorIO plugin is enabled, OpenRV does not support channel
selecting, reordering or dithering.
This PR allows the TopViewToolBar menus to function the same way it does
when OCIO is disabled. The main menu items work too for channel
reordering. The code added to the OCIOIPNode node is very similar to the
non-OCIO DisplayIPNode. The identical shader code is utilised for
dithering, and the extra properties are named the same.

```
color.dither
color.channelOrder
color.channelFlood
```


https://github.com/user-attachments/assets/b8962e1c-c0df-496e-bdc3-fcbb4b546846


The PR also includes 2 additional fixes
- Fixes a Qt issue where screen names end up being blank. I was running
into this on macOS where Qt is using deprecated apis to determine the
screen name.
- Fixes error message that can occur with recent changes to the OCIO
node. The fix temperraly disables the OCIONode during display/view
property changes. This Prevents it from rebuilding shaders while it may
be in an invalid state.

### Describe what you have tested and on which operating system.

Rocky 9, CentOS, MacOS

---------

Signed-off-by: Mark Reid <markreid@netflixanimation.com>