SG-39510: Block options for Live Review participants (#834)
### [SG-39510](https://jira.autodesk.com/browse/SG-39510): Block options
for Live Review participants

### Summarize your change.

Disable buttons, menus and related hotkeys that shouldn't be used by a
participant of a Live Review session. For buttons, their corresponding
tooltip is also updated to say "You must be presenter to use" when the
user is a participant. If the user is a presenter or if they are not in
a Live Review session anymore, all the disabled buttons and menus are
enabled again and the tooltips are set back to their default
description. When clicking on the media or using a hotkey, a widget is
displayed with the string "You are not the presenter". Moreover, the
Annotation panel will be closed when joining a Live Review session is it
open at that moment.

### Describe the reason for the change.

During a Live Review session, an RV user was still capable of clicking
on multiple buttons and menu options that could cause mismatch between
the presenter's screen at the participant screen. A Live Review
participant should only be allowed to click on options that don't change
the ongoing session.

### Describe what you have tested and on which operating system.

This is the list of disabled feature for a Participant of a Live Review
session:

- Control, Edit, Color, View, Source, Stack, Sequence and Layout menu
items
- Annotation, Session Manager and Default Views of the Tools menu
- All hotkeys for the above menu items
- Left-click and right-click on the timeline and the timeline magnifier
- All bottom toolbar options, except toggling image info, networking
dialog/Live Review panel, timeline, timeline magnifier, and volume
- All the top toolbar options, except toggling full-screen mode, framing
image in view, and selecting the background style

---------

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>