Fix 2x RemoteRvCommands (#847)
### Fix 2x RemoteRvCommands

### Linked issues
NA

### Summarize your change.

Note that this commit is only impacting Open RV when using the Open RV
Live Review functionality.

This commit adds addSourceVerbose() to the list of the RemoteRvCommands.

It also solves an issue with the addSourcesVerbose() command where the
RemoteRvCommands was not executed before it was defined after
before-progressive-loading which was preventing the generic-rv-command
event from being broadcasted to the participants.


### Describe the reason for the change.
addSourcesVerbose() and addSourceVerbose() were not behaving as expected
when used in RV Live Review.

### Describe what you have tested and on which operating system.
Successfully tested on macOS 

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>