Fix before/after-progressive-loading events imbalance (#854)
### Fix before/after-progressive-loading events imbalance

### Linked issues
NA

### Summarize your change.

This commit fixes a potential imbalance between
before-progressive-loading and after-progressive-loading events when
progressive source loading is active.
This commit only affects Open RV when the progressive source loading is
enabled
Which can be done via by setting the following env var:
export RV_PROGRESSIVE_SOURCE_LOADING=1

Also in this commit : properly initialized a progressive source loading
variable m_loadingID which was uninitialized.

### Describe the reason for the change.

This event imbalance proved to be detrimental to a recent mechanism that
was introduced in RV Live Review where some commands such as
addSourceVerbose() or clearSession() for examples, are executed on the
participant RVs as native commands instead of broadcasting the numerous
RV graph changes which are a side effect of such meta command.

### Describe what you have tested and on which operating system.
Successfully tested on macOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>