SG-39510: Close annotation panel when joining Live Review (#852)
### [SG-39510](https://jira.autodesk.com/browse/SG-39510): Close
annotation panel when joining Live Review

### Summarize your change.

When joining a Live Review session, the Annotation panel should be
closed if it was still open. Since a participant shouldn't be allowed to
annotate the media in a standard Live Review session, there is no reason
for the panel to be available.

### Describe the reason for the change.

In a previous PR, options and menus related to annotations were blocked
for participants of a Live Review session. However, if the Annotation
panel was opened when joining a session, it would stay that way. This
can be confusing to the participant since all other options are
disabled.

### Describe what you have tested and on which operating system.

Testing joining a Live Review session with the Annotation panel opened
and closed was perform on macOS

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>