Fix Window>Fit (w shortcut) for high dpi screen (#864)
### Fix Window>Fit (w shortcut) for high dpi screen

### Linked issues
n/a

### Summarize your change.
Take the DPI scaling into account when calculating the width and height
in the function.

There are one other issues that was discovered during testing:
Intermittent: Sometime the scaling of the UI get fuzzy when moving
OpenRV from a non-high-dpi screen to a high dpi screen.

### Describe the reason for the change.
Pressing W multiple time (or using the menu) would crash OpenRV after 3
times.

### Describe what you have tested and on which operating system.
MacOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

---------

Signed-off-by: Cédrik Fuoco <cedrik.fuoco@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>