SG-39830: Add refreshOutputVideoDevice to CommandsModule (#863)
### [SG-39830](https://jira.autodesk.com/browse/SG-39830): Add
refreshOutputVideoDevice to CommandsModule

### Summarize your change.

Add refreshOutputVideoDevice in CommandsModule.cpp and
rv_commands_setup.py to access the function in Python.

### Describe the reason for the change.

When loading a media file, RV calls systematically
`setOutputVideoDevice` in `readGTO` of RVSession.cpp. However, when
joining an OTIO-based Live Review session, this path is never used and
the output video device will not be set properly after quitting and
joining back a session. To replicate the normal behaviour the output
video device should be set when joining an OTIO-based Live Review
session. To do so, that C++ function needs to be available in Python
through the commands module. Note that this behaviour was not affecting
RV to RV Live Review session since in that case, `readGTO` was being
called.

### Describe what you have tested and on which operating system.

Quitting and joining back an OTIO-based Live Review session with the
presentation mode enabled through NDI in RV.

---------

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>