SG-39984: Fix duration and startFrame paint properties (#868)
### [SG-39984](https://jira.autodesk.com/browse/SG-39984): Fix duration
and startFrame paint properties

### Summarize your change.

Change fallback values of duration and startFrame when the properties
are not defined on the paint node. For duration, that value should have
been 1 instead of 0. For startFrame, the frame value is access through
the pen component that is format as follows:
`"pen:strokeId:frameNumber:user"`. Even if it wasn't causing an issue, I
changed the `ghostBefore` and `ghostAfter` values to 5 instead of 0 to
match what the SessionIPNode is setting since 0 doesn't make sense here.

### Describe the reason for the change.

The `duration` and `startFrame` properties were added to know when and
for how long to display held and ghosted annotations as well as
calculate the opacity level for ghosted annotations. However, these two
properties doesn't exist in previous version of RV. This is causing an
issue, because annotations saved in an older RV session won't have these
new properties and the hold and ghost feature will not be displayed
properly.The fallback values are now taking care of that case.

### Describe what you have tested and on which operating system.

Enabling hold and ghost on an RV session from RV 2024.2.1 in RV 2025.0.0
was tested on macOS.

---------

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>