Fix issue where the wrong QProcess.finished signal was used (#869)
### Fix issue where the wrong QProcess.finished signal was used

### Linked issues
n/a

### Summarize your change.
I added some logic to discard known deprecated signals. The logic is
present in MuQt5 and MuQt6 in prevision of future deprecated signals.

There is no reliable way to detect if a signal is deprecated. Therefore,
we have to rely on known lists.

### Describe the reason for the change.
In Mu code, there is an issue where the wrong QProcess.finished is used
because there is a deprecated version of it from Qt 5.13.

### Describe what you have tested and on which operating system.
MacOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Cédrik Fuoco <cedrik.fuoco@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>