SG-38881: Refactor Hold and Ghost OTIO schema (#871)
### [SG-38881](https://jira.autodesk.com/browse/SG-38881): Refactor Hold
and Ghost OTIO schema

### Summarize your change.

Move the `hold`, `ghost`, `ghost_before` and `ghost_after` properties
into the Annotation schema. The Annotations hook was also updated to
properly retrieve the properties with that change. Another PR will be
added in Commercial RV to update the corresponding OTIO reader tests.

### Describe the reason for the change.

The `hold` and `ghost` properties were sent by the web app through the
track's metadata and `ghost_before` and `ghost_after` were not sent in
the OTIO. After discussion, it has been agreed that all four properties
should be directly in the Annotation schema.

### Describe what you have tested and on which operating system.

Enabling and disabling Hold and Ghost in a web to RV Live Review session
was tested on macOS.

---------

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>