Fix fit window to image crash - w keyboard shortcut (#886)
### Fix fit window to image crash - w keyboard shortcut

### Linked issues
Fixes #885

### Describe the reason for the change.
**Problem:**
RV was sometimes crashing when fitting the window to image - w keyboard
shortcut

### Summarize your change.

**Cause:**
The source of this isssue is that the FitWindowToImage functionality
uses a new Qt6 functionality: QGuiApplication::screenAt() wrapped in the
Mu command qt.QGuiApplication.screenAt().
Now this new Qt6 method can possibly returns nullptr if the position nis
outside any visible screen.
The new associated Mu binding was always wrapping the returned pointer
as a Mu object no matter what: makeinstance<QScreenType>.
Which is why the crash occurred: when the position was outside the
screen then the Mu code would reference the nullptr.

**Solution:**
Now returning a nullptr for all Mu bindings wrapping a QScreen as a Mu
object: makeinstance<QScreenType>
I did that for all the places where we were wrapping a QScreen in Mu as
all the instances could actually be null as per the Qt6 documentation.

### Describe what you have tested and on which operating system.
Successfully tested on macOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>