SG-39966: Fix annotation color selection (#877)
### [SG-39966](https://jira.autodesk.com/browse/SG-39966): Fix
annotation color selection

### Summarize your change.

Replace QColorDialog `currentColorChanged` signal with `colorSelected`.

### Describe the reason for the change.

`currentColorChanged` is a signal emitted whenever the color is changed
in the dialog. This means that "Ok" and "Cancel" had no effect since the
color was already changed as soon as it was selected. `colorSelected` is
emitted just after "Ok" is clicked to set the color to use. That way,
the color is only set after the user has clicked on "Ok", quitting the
dialog or clicking on "Cancel" will not save the selected color.

### Describe what you have tested and on which operating system.

Selecting colors through the QColorDialog was tested on macOS.

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>