Format Python code with Ruff (#936)
### Format Python code with Ruff

### Summarize your change.

Run `ruff format` on all Python files with the default formatting rules.

The formatting was done following the default rules but the line length
was increased to 120 in `ruff.toml` that is added in another PR.

### Describe the reason for the change.

Ruff is a tool that can do exactly what Black was doing as a formatter
while also being able to lint and fix the code following PEP8 at the
same time. Using the same tool for both makes the steps of linting and
formatting easier to run locally, with pre-commit and on the CI.

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>