SG-40626: Garbage returned by Mu's exception handling if returning value inside catch block (#941)
### Linked issues

SG-40626


### Describe the reason for the change.

Under certain circumstnaces, returning a value from within a mu catch
block returned garbage data.

### Summarize your change

This is one of those changes that requires some background and a
thorough explanation before describing the fix.... so here goes!

While testing a PR for menu refactoring and event categorization, I
noticed that when RV started (without loaded media), that certain menu
items are enabled when they shouldn't be. For example, Flip/Flop should
have been disabled, but they were not! But, my code seemed to be
correct.

<img width="1388" height="838" alt="image"
src="https://github.com/user-attachments/assets/17040c61-3a58-41f0-95df-52c7f976e676"
/>

For these items, the state callback passed as a validator to the menu
item is declared as an alias to a function that returns a lambda, in
rvui.mu:

_toggleFlipState = toggleIntPropState("#RVTransform2D.transform.flip")_

<img width="1078" height="148" alt="image"
src="https://github.com/user-attachments/assets/0ff78289-7778-42a8-a58c-0c8f17ade2f0"
/>


In our PR that exhibits the bug, we take the state func
(toggleFlipState) and combine it with another category-validating
function to create a composite function that tests both the category and
the state. This composite function is called by the menu item handler to
determine if the item should be enabled/disabled. It's when using this
composite function that the bug manifested itself.

The bug showed up
when _toggleIntPropState()_ calls getIntProperty(...) inside a Mu
try block. If the property doesn't exist (which is the case when no
media is loaded at RV startup), the C++ code for getIntProperty in
CommandsModule.cpp throws an exception, which is caught by the Mu catch
block, and should return DisabledMenuState (aka: -1) from within the Mu
catch block.

That should have worked just fine (and it did work just before my
refactoring), but it didn't anymore after I refactored: our menu item
stayed enabled. My composite function looked fine, too, so I had to dig
deeper.

To debug this, I added traces in toggleIntPropState() to see when if
enter the function, if we enter the catch, if we return the correct
value, and if anything executes after the return in the Mu catch block.
I also printed the returned value in the composite function that used
the propStateFunc.

<img width="1214" height="1008" alt="image"
src="https://github.com/user-attachments/assets/4d670968-1123-490f-8424-acf42ec8ceae"
/>


and

<img width="1228" height="650" alt="image"
src="https://github.com/user-attachments/assets/e1f516a9-5691-479f-a7b9-1b0828d1c54a"
/>

And we got interesting results: 

First, we saw the traces appear when we entered the propStateFunc, and
when we entered the catch block, as expected. Thus, the code did indeed
"return DisabledMenuState" (-1) from within the catch block.   BUT the
composite function that consumed the return value from said
propStateFunc showed a garbage value (e.g., 187387624), which is
definitely not the expected -1.

<img width="930" height="238" alt="image"
src="https://github.com/user-attachments/assets/f76c9cf7-6765-447c-adee-06a2a1655065"
/>


To confirm that the problem only existed when returning from within a
catch block, I just moved the return statement to just outside the catch
block, and lo and behold, the problem was fixed! (Flip/Flop menu items
were now disabled, as expected)

Conclusion: There's a problem in Mu's exception handling when
returning from inside a catch block.

=================

Further problem analysis/debugging/solving.

When Mu calls getIntProperty() (in CommandsModule.cpp) and the property
doesn't exist, the C++ code throws an exception
via throwBadPropertyType(), which is supposed to be caught by
the catch block in Mu. The C++ catch code for mu is implemented in
ExceptionType::mu__catch(), which contains some code that looks like
black magic invocations, but I digress.

Anyway, after a lot of digging, the problem is that in
Mu, return statements are implemented via longjmp, while try/catch is
implemented via C++ exceptions. When a return occurs inside
a catch block, the longjmp must traverse C++ exception handling code,
which is not allowed (in fact it's undefined behaviour as per the C++
standard).

Indeed, in ExceptionType::mu__catch, there was simply no mechanism
whatsoever to properly preserve the return value when doing return from
the catch block; and the longjmp call does not appear to be set up in a
way that it traverses the C++ exception handling code safely. Its
actually unclear to me why this code used to work at all, I think it
might have been because the stack was left in an "undisturbed" state, or
at least, still in a state where the return value happened to remain at
the right place in memory, but I don't know for sure.

Anyway, to fix this, we need to modify
mu__catch() and mu__catch_all() to Install a setjmp that
captures return statements from the catch body, and if
a return is detected, throws an internal C++ exception that carries the
returned value with it. Then, mu__try() catches this exception, and
converts it to a proper longjmp _after_ the exception was handled, which
allows the function to exit properly with the returned value (whew, holy
code path, batman!).



### Describe what you have tested and on which operating system.

macOS

### Add a list of changes, and note any that might need special
attention during the review.

Due to the very unusual type of fix this is, I don't expect anyone to be
able to look at this code change and go "LGTM!" right away. In fact, I
expect people to be more like : "what the hell am I even looking at?".

For this reason, it's probably the only time I will suggest that code
reviewers who look at this actually pull the branch locally, and ask AI
the following:

_"What are these changes about, why are they necessary, what do the
changes do, how does it work, is the fix correct and if so, why is it
correct, does it work with all return types, even complex types?"_

### If possible, provide screenshots.

---------

Signed-off-by: Patrick Bergeron <patrick.bergeron@autodesk.com>