added bind to event "clear-annotations-current-frame" (#929)
### Linked issues

### Summarize your change.

Added two events that can be sent internally, to clear the current frame
of annotations. I want to call this internal event from rv_tests.py,
when we auto-draw-annotations (circles).

Currently, adding those two bindings changes nothing to the code, since
it's not even used outside the context of rv_tests.py

### Describe the reason for the change.

This is necessary for auto-draw-annotations in rv_test.py, where I'd
like to clear the frame every 10 annotations (otherwise, in stress
tests, annotations keep acumulating, and the render gets slower and
slower and slower and slower.

### Describe what you have tested and on which operating system.

rv_tests.py, which is in Commercial RV. 

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Patrick Bergeron <patrick.bergeron@autodesk.com>